Condition variable timeouts, suspend folded into sleep state, - add capability to resume with return value - add std::chrono! - huge fix in fast voluntary context switch! - optimize ISR code - implement Indirect Branch Prediction Barrier spectre mitigation - up to C++17 from C++14 - Added __DGOS__ define, defined in all bootloader variations too - improve virtio config - add spice - use .balign instead of .align - do instruction fixups up front and not atomically - fix issue in bootloader not loading large files correctly - automatic breakpoint in vpanic - remove sub_rsp, change add_rsp to adj_rsp - implement std::gcd and std::lcm - improve ECAM init - add way to reenable IRQ in cpu_scoped_irq_disable - Use instruction patching for xsaves/xsaveopt/xsavec/xsave/xrstor/xrstors and the 64 versions of the same instead of macro duplication and patched jump - Partial UB fix in GDT init (union misuse) - Improve GDT class - Drastically improve xsave init - Fix cpu mask - Synchronization code cleanup - Add missing clobber in critical section entry asm wrapper - Fix illegal use of 0x66 to pad vzeroall, it is documented to be unpredictable - Fix the bottom of call stacks so ?? entry is not there anymore - Fix CPU affinity bugs in large mask - Use C++ locks/condition variables instead of raw low level ones in IDE driver - Add wait_until to iocp - Fix PCI address bugs - Reimplement PCI device caching - Clean up UART driver to not misuse unions and cause UB. use memcpy! - Verify endpoint removal in XHCI driver - Found bug in QEMU, caused major problems with IRQs on virtio devices - Fix adding initrd TRAILER!!! - Recreate kernel.config (preprocessor defines for syntax highlight) - Add std::ratio - Add several type_traits - Reduce duration of render test - Fix some allocation function attributes - Add way to efficiently notify N waiters on condition variable - Fix operator new declaration - Remove unnecessarily volatile in lock-protected condition variable - Partial futex - Up to GCC 9.1.0 - Make sure RTL8139 is present before loading module - Sorry, too much for one commit I know

+++ boot/paging.h
@@ -107,7 +107,7 @@ struct iovec_t {
     uint64_t size;
 };
 
-int paging_iovec(iovec_t **ret, uint64_t vaddr,
+size_t paging_iovec(iovec_t **ret, uint64_t vaddr,
                  uint64_t size, uint64_t max_chunk);
 
 off_t paging_iovec_read(int fd, off_t file_offset, uint64_t vaddr,
